[5.x] Token class changes (#9964)
