+++ src/Tokens/TokenRepository.php
@@ -2,58 +2,15 @@
 
 namespace Statamic\Tokens;
 
-use Illuminate\Support\Carbon;
-use Statamic\Facades\File;
-use Statamic\Facades\YAML;
+use Statamic\Contracts\Tokens\Token as TokenContract;
+use Statamic\Contracts\Tokens\TokenRepository as Contract;
 
-class TokenRepository
+abstract class TokenRepository implements Contract
 {
-    public function make(?string $token, string $handler, array $data = []): Token
+    public function make(?string $token, string $handler, array $data = []): TokenContract
     {
-        return new Token($token, $handler, $data);
+        return app()->makeWith(TokenContract::class, compact('token', 'handler', 'data'));
     }
 
-    public function find(string $token)
-    {
-        $path = storage_path('statamic/tokens/'.$token.'.yaml');
-
-        if (! File::exists($path)) {
-            return null;
-        }
-
-        return $this->makeFromPath($path);
-    }
-
-    public function save(Token $token)
-    {
-        File::put(storage_path('statamic/tokens/'.$token->token().'.yaml'), $token->fileContents());
-
-        return true;
-    }
-
-    public function delete(Token $token)
-    {
-        File::delete(storage_path('statamic/tokens/'.$token->token().'.yaml'));
-
-        return true;
-    }
-
-    public function collectGarbage()
-    {
-        File::getFilesByType(storage_path('statamic/tokens'), 'yaml')
-            ->map(fn ($path) => $this->makeFromPath($path))
-            ->filter->hasExpired()
-            ->each->delete();
-    }
-
-    private function makeFromPath(string $path): Token
-    {
-        $yaml = YAML::file($path)->parse();
-
-        $token = basename($path, '.yaml');
-
-        return $this
-            ->make($token, $yaml['handler'], $yaml['data'] ?? [])
-            ->expireAt(Carbon::createFromTimestamp($yaml['expires_at']));
-    }
+    abstract public static function bindings(): array;
 }
